Code refactoring
Separating definitions and functions
Cleaning up the codebase
